backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
sse-starlette>=2.0.0
openai==1.101.0
pymupdf==1.26.4
PyPDF2==3.0.1
python-dotenv==1.1.0
tiktoken==0.11.0
pyyaml==6.0.2
anthropic>=0.40.0
httpx>=0.27.0
google-genai>=1.0.0
supabase>=2.0.0
PyJWT>=2.8.0
